Working on the Assignment Operator.
Seeing as our scale production doesn't return the actual scale integer, but just that it is a scale, I'll be skipping 2-halves.fscalr. I'm currently working on getting misc.fscalr to work, in so far that we have a function that will return a sequenve by a simple call to \'evaluate(params).\'

I did actually modify SimpleNode.java, so let me un-ignore it.