Modified ScalrParser.jjt.
So, I'm still trying to figure out how to evaluate this tree. To that extent, I added some JJTree productions to ScalrParser.jjt so that I know if the operation is addition, subtraction, multiplication, division, or modulous, since that information doesn't seem to by any of the other nodes in that tree, unlike with assignment where that's given by the assign node.
The PlusOperator class will be changed to the BinaryOperator class, which may be confusing as Assignment is also a binary operator, but it's different enough to warrant its own class. The binary operators, \+, \-, \/, \*, \%, are all similar enough to be lumped together.