It all of a sudden made sense.
Despite how wierd the tree looks, it's pretty regular, which means recursive walking shouldn't be too difficult. The number of functions to achieve this is skyrocketing, but it can't be helped with each production. I feel like writing my own tree printer that also prints the value of each tree.