Committing changes for tonight.
Function calling and function declaration with a variable number of parameters is now pretty much set. I am still trying to sort through when expressions should and should not contain variable references, but the answer to that seems to be whenever we see an ID and we are not in assignment "=".