Commented WhileStatement.java.
Modifying IfElseStatement.java to follow WhileStatement.java pretty closely. I was able to design IfEsleStatement so that adding statements to the appropriate if/else if/else was just as transparent as adding while statements. Although the grammar should handle this, the class is designed to prevent the addition of an else if after adding an else.