Fixed parameter calling in Functino.java.
It now adds the passed in expressions into the symbol table in the same order as the parameter name declaration. If, for some reason, the number of passed in parameters is not the same as was expected, it prints out to the error stream as such, as well as the name of the arguments it was expecting. Also fixed the toString method of note to print the correct durations.