Fixed the getType() method of VariableReference.
The fix was adding a new table to SymbolTable that contains the type of CreationOperators. Since a variable is not added to SymbolTable until the getValue of CreationOperator is executed, we could not get its type. By adding this table, creation operator adds the type of the rval before it is executed, since we know what it is. VariableReference new refers to that new table for getType().