Function calling works.
My grammar doesn't change scope yet, so an infinite loop was occuring as it was trying to get the reference to a formal parameter which was an argument which was a formal parameter which was an argument etc. Now, the only thing left is import statements, tscalr productions, fscalr productions, and tracks. I'll do tracks, and go to bed.