Added a pitch() and duration() operator with no extra conflicts, so nothing should have broken.
The pitch operator takes a note as its argument, and returns the ordinal of the pitch, not the midivalue.
The pitch operator takes a note or sequence as its argument, and returns the ordinal of the length, if the argument is a note, or the number of notes in a sequence.