Found the problem.
The fundamental problem is that expressions aren't copied when functions are copied. That is, when a function is called, the expressions it evaluates are globally the same. That means that sequences, who modify themselves when evaluated, are shared with all function calls. Thus, the notes we see in testfile.tscalr aren't changed because it is the exact same sequence that is being added. This means that the getValue method of each of the Variables should return a copy of themselves, or should they?